Message port close `events` channel after all the sender routines exited
When the context of the `Listen()` being canceled. There is a race condition that the goroutine that closes the `events` channel runs prior than the other handler goroutines (for both `message`/`messageerror`). If this happens, those handler goroutines will panic as the `events` channel is closed, and sending to a closed channel just panics in Go.